\name{adjustedTableLaTeX}
\alias{adjustedTableLaTeX}
\title{Creates table LaTeX table of proportions for the two given variables. The
resulting table has two columns for each unique value in y, one corresponding
to the overall row column proportions and the second adjusted proportion (i.e.
proportion excluding missing values).}
\usage{
  adjustedTableLaTeX(x, y, caption = "", label = "", ...)
}
\description{
  Creates table LaTeX table of proportions for the two
  given variables. The resulting table has two columns for
  each unique value in y, one corresponding to the overall
  row column proportions and the second adjusted proportion
  (i.e. proportion excluding missing values).
}

