\name{cohortSummaryLaTeX}
\alias{cohortSummaryLaTeX}
\title{Prints LaTeX of cohort summary.}
\usage{
  cohortSummaryLaTeX(ret, category = "Category",
    category.align = "l", caption = "Retention Summary",
    label = "retentionSummary",
    sparkline.retention.rect = NULL,
    sparkline.completion.rect = NULL, sparkline.width = 10,
    sparkline.height = NULL, completion.month = 48)
}
\description{
  Prints LaTeX of cohort summary.
}

