\name{createLocalRepos}
\alias{createLocalRepos}
\title{This package will download all packages for the given type to \code{local.repos}
from the given repository.}
\usage{
  createLocalRepos(local.repos,
    repos = "http://cran.r-project.org", type = "source")
}
\description{
  This package will download all packages for the given
  type to \code{local.repos} from the given repository.
}
\seealso{
  \code{\link{download.packages}} which this function wraps
}

