\name{getAgeGroups}
\alias{getAgeGroups}
\title{Returns a factor variable using the given breaks and labels. Will calculate the
age based upon the calcDate if given, otherwise the ageCol will be used.}
\usage{
  getAgeGroups(dateCol, calcDate, breaks, labels,
    ageCol = NULL)
}
\description{
  Returns a factor variable using the given breaks and
  labels. Will calculate the age based upon the calcDate if
  given, otherwise the ageCol will be used.
}
\author{
  Jason Bryer <jbryer@excelsior.edu>
}

