\docType{package}
\name{itutils-package}
\alias{irutils}
\alias{itutils-package}
\title{Utility functions for Institutional Research}
\description{
  Utilities for Intitutional Research. This package is
  designed primarly for institutional researchers, though
  many of the functions provided are appropriate for a
  broader audience. Many of the functions are documented in
  the (forthcoming) manual, Introduction to R and LaTeX for
  Institutional Research.
}
\author{
  \email{jason@bryer.org}
}
\keyword{institutional}
\keyword{package}
\keyword{research}

