\name{local.available.packages}
\alias{local.available.packages}
\title{Returns a list of available packages in the local repository.}
\usage{
  local.available.packages(local.repos, type = "source")
}
\description{
  Returns a list of available packages in the local
  repository.
}
\seealso{
  \code{\link{available.packages}} which this function
  wraps
}

