\name{local.install.packages}
\alias{local.install.packages}
\title{Installs packages from the local repository.}
\usage{
  local.install.packages(pkgs, local.repos, type =
  "source")
}
\description{
  Installs packages from the local repository.
}
\seealso{
  \code{\link{install.packages}} which this function wraps
}

