\name{lsos}
\alias{lsos}
\title{An enhanced version of ls.}
\usage{
  lsos(..., n = 10)
}
\arguments{
  \item{...}{unspecified parameters}

  \item{n}{the (largest) number of items to list}
}
\description{
  An enhanced version of ls.
}
\author{
  http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}

