\name{pisa}
\alias{pisa}
\title{United States results from the 2009 Programme of International Student Assessment (PISA)
as provided by the Organization for Economic Co-operation and Development (OECD).
See \url{http://www.pisa.oecd.org/} for more information including the code book.}
\format{a data frame 5,233 ovservations of 437 variables.}
\description{
  United States results from the 2009 Programme of
  International Student Assessment (PISA) as provided by
  the Organization for Economic Co-operation and
  Development (OECD). See \url{http://www.pisa.oecd.org/}
  for more information including the code book.
}

