\name{plot.likert.heat}
\alias{plot.likert.heat}
\title{Function creates and returns a tabular heatmap ideal for summarizing likert
survey items.}
\usage{
  plot.likert.heat(likert, low.color = "white",
    high.color = "blue", text.size = 2)
}
\arguments{
  \item{items}{the items to summarize.}
}
\description{
  Function creates and returns a tabular heatmap ideal for
  summarizing likert survey items.
}

