\name{plotHeatmapTable}
\alias{plotHeatmapTable}
\title{Function creates and returns a tabular heatmap ideal for summarizing likert
survey items.}
\usage{
  plotHeatmapTable(items, nlevels = length(levels(items[,
  1])), low.color = "white", high.color = "blue", text.size
  = 2)
}
\arguments{
  \item{items}{the items to summarize.}
}
\description{
  Function creates and returns a tabular heatmap ideal for
  summarizing likert survey items.
}

