\name{retentionSummaryLaTeX}
\alias{retentionSummaryLaTeX}
\title{Prints LaTeX of retention summary.}
\usage{
  retentionSummaryLaTeX(ret, category = "Category",
    category.align = "l", caption = "Retention Summary",
    label = "retentionSummary")
}
\description{
  Prints LaTeX of retention summary.
}

