\name{updateLocalRepos}
\alias{updateLocalRepos}
\title{This function will update the local repository with new packages and/or newer versions.}
\usage{
  updateLocalRepos(local.repos, repos =
  "http://cran.r-project.org", type = "source")
}
\description{
  This function will update the local repository with new
  packages and/or newer versions.
}

