\docType{package}
\name{itutils-package}
\alias{irutils}
\alias{itutils-package}
\title{Utility functions for Institutional Research}
\description{
  Database Access Utility Functions
}
\author{
  \email{jason@bryer.org}
}
\keyword{institutional}
\keyword{package}
\keyword{research}

