\name{plotBarchartTable}
\alias{plotBarchartTable}
\title{Creates a bar chart using percentages ideal for likert items.}
\usage{
  plotBarchartTable(items, grouping = NULL, nlevels =
  length(levels(items[, 1])), low.color = "blue",
  high.color = "red", neutral.color = "white", text.size =
  2)
}
\arguments{
  \item{items}{the items to summarize.}
}
\description{
  Creates a bar chart using percentages ideal for likert
  items.
}

